# resBot/swagger.yaml
swagger: "2.0"
info:
  description: "This is a brief description of Resala bot API, The API provides ability to add new members, new contributions and query contributions of a membe. You can see the database design at https://dbdiagram.io/d/5e24a71a9e76504e0ef06cf2 "
  version: "1.0.0"
  title: "Resala Bot API"
  termsOfService: "http://swagger.io/terms/"
  contact:
    email: "aaahussein@gmail.com"
  license:
    name: "Apache 2.0"
    url: "http://www.apache.org/licenses/LICENSE-2.0.html"
host: "resala.org.bot" #To be changed
basePath: "/v1"
...
schemes:
- "https"
- "http"
consumes:
      - "application/json"
produces:
      - "application/json"
paths:
...
  /bot/{bot_id}/member/{phone}/contribution:
    post:
      tags:
      - "contribution"
      summary: "add a new contribution to by this member belonging to this bot"
      parameters:
      - name: "bot_id"
        in: "path"
        description: "The id of the bot you want to add a member to it"
        required: true
        type: "string"
        format: "uuid"
      - in: "path"
        name: "phone"
        description: "The phone of the member that we want to add a contribution to him"
        required: true
        type: "string"
      - in: "body"
        name: "contribution"
        schema:
          $ref: "#/definitions/Contribution"
      responses:
        201:
          description: "Added successfully"
        400:
          description: "Missing required data "
        404:
          description: "Bot or member not found"
    get:
        tags:
        - "contribution"
        summary: "get all contributions of that member in a given time range defined by start_date and end_date"
        parameters:
        - name: "bot_id"
          in: "path"
          description: "The id of the bot you want to add a member to it"
          required: true
          type: "string"
          format: "uuid"
        - in: "path"
          name: "phone"
          description: "The phone of the member that we want to get his contributions"
          required: true
          type: "string"
        - in: "query"
          name: "start_date"
          type: "string"
          format: "date"
        - in: "query"
          name: "end_date"
          type: "string"
          format: "date"
        responses:
          200:
            description: "get all contributions whose date falls between startDate(inclusive) and endDate(inclusive)"
            schema:
              type: "array"
              items:
                $ref: "#/definitions/Contribution"
          404:
            description: "Bot or member not found"
...
definitions:
...
  Contribution:
    type: "object"
    properties:
      contribution_type:
        type: "string"
      date:
        type: "string"
      details:
        type: "string"
      amount_money:
        type: "integer"
        format: "int64"
      did_receive_money:
        type: "boolean"
    required:
      - "contribution_type"
    xml:
      name: "Contribution"
...